+++ Assets/PatchKit/UnityEditorExtension/Editor/Core/AppBuild.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Linq;
 using JetBrains.Annotations;
 using UnityEditor;
+using UnityEngine;
 #if UNITY_2018_1_OR_NEWER
 using UnityEditor.Build.Reporting;
 #endif
@@ -58,14 +59,8 @@ public static class AppBuild
     }
 
     [NotNull]
-    private static readonly string[] WindowsPdbFiles =
-    {
-        "player_win_x86.pdb",
-        "player_win_x86_s.pdb",
-        "player_win_x64.pdb",
-        "player_win_x64_s.pdb"
-    };
-
+    public static IEnumerable<string> WarningFiles = Array.Empty<string>();
+    
     public static string Location
     {
         get
@@ -75,12 +70,12 @@ public static class AppBuild
 
             string validationError = GetLocationValidationError(value);
 
-            if (validationError == null)
+            if (validationError != null)
             {
-                return value;
+                Debug.LogError(validationError);
             }
 
-            return null;
+            return value;
         }
         set
         {
@@ -97,18 +92,19 @@ public static class AppBuild
         }
     }
 
-    private static bool AreThereOnlyBuildEntries(
+    [NotNull]
+    private static IEnumerable<string> GetFilesOutsideOfBuildEntries(
         [NotNull] string location,
         params string[] buildFiles)
     {
         if (location == null)
         {
-            throw new ArgumentNullException("location");
+            throw new ArgumentNullException(nameof(location));
         }
 
         if (buildFiles == null)
         {
-            throw new ArgumentNullException("buildFiles");
+            throw new ArgumentNullException(nameof(buildFiles));
         }
 
         string parentDirPath = Path.GetDirectoryName(location);
@@ -117,28 +113,64 @@ public static class AppBuild
 
         string[] entries = Directory.GetFileSystemEntries(parentDirPath, "*");
 
-        return entries.All(
+        return entries.Where(
             x =>
             {
                 string fileName = Path.GetFileName(x);
 
                 if (Path.GetDirectoryName(x) != parentDirPath)
                 {
-                    return true;
+                    return false;
                 }
 
                 if (buildFiles.Contains(fileName))
                 {
-                    return true;
+                    return false;
                 }
 
-                if (Platform.IsWindows() && WindowsPdbFiles.Contains(fileName))
+                if (Platform.IsWindows() && fileName.ToLower().EndsWith(".pdb"))
                 {
-                    return true;
+                    return false;
                 }
 
-                return false;
+                return true;
             });
+    }
+
+    
+    [NotNull]
+    private static IEnumerable<string> GetLocationValidationWarningFiles(string location)
+    {
+        switch (Platform)
+        {
+            case AppPlatform.Windows32:
+            case AppPlatform.Windows64:
+                string winBuildFileName = Path.GetFileName(location);
+                string winBuildDirName =
+                    winBuildFileName.Replace(".exe", "_Data");
+
+                return GetFilesOutsideOfBuildEntries(
+                    location,
+                    winBuildFileName,
+                    winBuildDirName,
+                    "MonoBleedingEdge",
+                    "Mono",
+                    "UnityCrashHandler32.exe",
+                    "UnityCrashHandler64.exe",
+                    "UnityPlayer.dll",
+                    "WinPixEventRuntime.dll");
+            case AppPlatform.Linux32:
+            case AppPlatform.Linux64:
+                break;
+            case AppPlatform.Mac64:
+                break;
+            case null:
+                break;
+            default:
+                throw new ArgumentOutOfRangeException();
+        }
+
+        return Array.Empty<string>();
     }
 
     private static string GetLocationValidationError(string location)
@@ -163,24 +195,6 @@ public static class AppBuild
                         "Invalid build location file extension. Should be .exe.";
                 }
 
-                string winBuildFileName = Path.GetFileName(location);
-                string winBuildDirName =
-                    winBuildFileName.Replace(".exe", "_Data");
-
-                if (!AreThereOnlyBuildEntries(
-                    location,
-                    winBuildFileName,
-                    winBuildDirName,
-                    "MonoBleedingEdge",
-                    "Mono",
-                    "UnityCrashHandler32.exe",
-                    "UnityCrashHandler64.exe",
-                    "UnityPlayer.dll",
-                    "WinPixEventRuntime.dll"))
-                {
-                    return "Build location must be an empty directory.";
-                }
-
                 break;
             case AppPlatform.Linux32:
             case AppPlatform.Linux64:
@@ -194,7 +208,7 @@ public static class AppBuild
 
                 string macBuildFileName = Path.GetFileName(location);
 
-                if (!AreThereOnlyBuildEntries(location, macBuildFileName))
+                if (GetFilesOutsideOfBuildEntries(location, macBuildFileName).Any())
                 {
                     return "Build location must be an empty directory.";
                 }
@@ -209,7 +223,7 @@ public static class AppBuild
         return null;
     }
 
-    public static bool TryCreate()
+    public static bool TryCreate(bool removePdbFiles)
     {
         bool success;
 #if UNITY_2018_1_OR_NEWER
@@ -231,12 +245,23 @@ public static class AppBuild
                 BuildOptions.None));
 #endif
 
-        if (success)
+        if (!success)
+        {
+            return false;
+        }
+
+        WarningFiles = GetLocationValidationWarningFiles(Location);
+        foreach (string warningFile in WarningFiles)
+        {
+            Debug.LogWarning("Unknown file in build location: " + warningFile);
+        }
+            
+        if (removePdbFiles)
         {
             RemovePdbFiles();
         }
 
-        return success;
+        return true;
     }
 
     private static void RemovePdbFiles()
@@ -247,14 +272,13 @@ public static class AppBuild
             case AppPlatform.Windows64:
                 string parentDirPath = Path.GetDirectoryName(Location);
                 Assert.IsNotNull(parentDirPath);
-
-                foreach (string pdbFile in WindowsPdbFiles)
+                string[] entries = Directory.GetFileSystemEntries(parentDirPath, "*");
+                
+                foreach (string pdbFile in entries.Where(e => e.ToLower().EndsWith(".pdb")))
                 {
-                    string pdbFilePath = Path.Combine(parentDirPath, pdbFile);
-
-                    if (File.Exists(pdbFilePath))
+                    if (File.Exists(pdbFile))
                     {
-                        File.Delete(pdbFilePath);
+                        File.Delete(pdbFile);
                     }
                 }
 
@@ -313,6 +337,7 @@ public static class AppBuild
                 if (validationError == null)
                 {
                     Location = location;
+                    WarningFiles = GetLocationValidationWarningFiles(location);
                 }
                 else
                 {