+++ Assets/PatchKit/UnityEditorExtension/Editor/Core/ExtensionVersion.cs
@@ -8,7 +8,7 @@ public static class ExtensionVersion
 
     public static int Minor = 0;
 
-    public static int Patch = 5;
+    public static int Patch = 6;
 
     [NotNull]
     public static string Name