+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.0.6]
 ### Fixed
 - Building when copy PDB files is selected in Build Settings
+- Building when files outside of build entries
 
 ### Added
 - New building option "Include .pdb files"
+- Warning about files outside of build entries 
 
 ## [1.0.5]
 ### Fixed