+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.6]
+### Added
+- New building option "Include .pdb files"
+- Warning about files outside of build entries
+
+### Fixed
+- Building when copy PDB files is selected in Build Settings
+- Building when files outside of build entries
+
 ## [1.0.5]
 ### Fixed
 - Reading API response body with invalid encoding